+++ srcs/render/utils.c
@@ -6,7 +6,7 @@
 /*   By: qmorinea <qmorinea@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 22:27:58 by qmorinea          #+#    #+#             */
-/*   Updated: 2025/04/08 22:39:41 by qmorinea         ###   ########.fr       */
+/*   Updated: 2025/04/09 10:57:59 by qmorinea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	put_pixel(t_mlx mlx, int x, int y, int color)
 	*(unsigned int *)dst = color;
 }
 
-float	to_radians(float degree)
+double	to_radians(double degree)
 {
 	return (degree * M_PI / 180);
 }
